+++ src/index.ts
@@ -6,4 +6,9 @@ export { Reader, ReaderModel, WebServiceClient };
 export * from './records.js';
 export * from './models/index.js';
 export * from './errors.js';
-export type { WebServiceClientError } from './types.js';
+export type {
+  ClientErrorCode,
+  WebServiceClientError,
+  WebServiceErrorCode,
+} from './types.js';
+export type { WebServiceClientOptions } from './webServiceClient.js';